Did some code restructuring
