+++ app/src/main/java/com/iyanuoluwa/imovie/MainActivity.kt
@@ -2,15 +2,13 @@ package com.iyanuoluwa.imovie
 
 import android.annotation.SuppressLint
 import android.os.Bundle
-import android.view.MotionEvent
-import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentPagerAdapter
 import androidx.viewpager.widget.ViewPager
 import com.google.android.material.bottomnavigation.BottomNavigationView
-import com.iyanuoluwa.imovie.custom.NaughtyPager
+import com.iyanuoluwa.imovie.util.custom.NaughtyPager
 
 class MainActivity : AppCompatActivity() {
