--- app/src/main/java/com/iyanuoluwa/imovie/api/Dates.kt
@@ -1,9 +1,0 @@
-package com.iyanuoluwa.imovie.api
-
-
-import com.google.gson.annotations.SerializedName
-
-data class Dates(
-    val maximum: String,
-    val minimum: String
-)
