--- app/src/main/java/com/iyanuoluwa/imovie/data/ApiPopular.kt
@@ -1,17 +1,0 @@
-package com.iyanuoluwa.imovie.data
-
-import com.iyanuoluwa.imovie.api.MovieJson
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-
-// https://api.themoviedb.org/3/movie/popular?api_key=68f286331e8795bd4addf043c1e8423d&language=en-US
-
-interface ApiPopular {
-    @GET("movie/popular?api_key=68f286331e8795bd4addf043c1e8423d&language=en-US")
-    fun getPopularMovies(
-            @Query("page") page : Int,
-            @Query("limit") limit : Int
-    ) : Call<MovieJson>
-}
