+++ app/src/main/java/com/iyanuoluwa/imovie/data/model/buffer/api/Crew.kt
@@ -1,4 +1,4 @@
-package com.iyanuoluwa.imovie.api2
+package com.iyanuoluwa.imovie.data.model.buffer.api
 
 
 import com.google.gson.annotations.SerializedName
