+++ app/src/main/java/com/iyanuoluwa/imovie/data/model/buffer/api/Dates.kt
@@ -1,0 +1,7 @@
+package com.iyanuoluwa.imovie.data.model.buffer.api
+
+
+data class Dates(
+    val maximum: String,
+    val minimum: String
+)
