+++ app/src/main/java/com/iyanuoluwa/imovie/ui/adapter/MovieAdapter.kt
@@ -1,4 +1,4 @@
-package com.iyanuoluwa.imovie.adapter
+package com.iyanuoluwa.imovie.ui.adapter
 
 import android.content.Context
 import android.content.Intent
